Problem04 is almost completed!

+++ oj-midterm/src/Problem04.java
@@ -2,10 +2,10 @@ import java.util.*;
 public class Problem04 {
     public static void main(String[] args) {
         Scanner input = new Scanner(System.in);
-        float h = input.nextInt();
-        float b = input.nextInt();
+        double h = input.nextInt();
+        double b = input.nextInt();
 
-       float a = (b*h)/2;
+       double a = (b*h)/2;
 
         System.out.println(a);
